+++ mmdet3d/datasets/sunrgbd_dataset.py
@@ -29,11 +29,12 @@ class SUNRGBDDataset(Custom3DDataset):
         # Use index to get the annos, thus the evalhook could also use this api
         info = self.data_infos[index]
         if info['annos']['gt_num'] != 0:
-            gt_bboxes_3d = info['annos']['gt_boxes_upright_depth']  # k, 6
-            gt_labels_3d = info['annos']['class']
+            gt_bboxes_3d = info['annos']['gt_boxes_upright_depth'].astype(
+                np.float32)  # k, 6
+            gt_labels_3d = info['annos']['class'].astype(np.long)
         else:
             gt_bboxes_3d = np.zeros((0, 7), dtype=np.float32)
-            gt_labels_3d = np.zeros(0, )
+            gt_labels_3d = np.zeros((0, ), dtype=np.long)
 
         anns_results = dict(
             gt_bboxes_3d=gt_bboxes_3d, gt_labels_3d=gt_labels_3d)
